move plugin provenance to McpManager
