+++ codex-rs/core/src/plugins/manager.rs
@@ -100,33 +100,10 @@ impl PluginCapabilitySummary {
     }
 }
 
-#[derive(Debug, Clone, Default, PartialEq, Eq)]
-pub struct ToolPluginProvenance {
-    plugin_display_names_by_connector_id: HashMap<String, Vec<String>>,
-    plugin_display_names_by_mcp_server_name: HashMap<String, Vec<String>>,
-}
-
-impl ToolPluginProvenance {
-    pub fn plugin_display_names_for_connector_id(&self, connector_id: &str) -> &[String] {
-        self.plugin_display_names_by_connector_id
-            .get(connector_id)
-            .map(Vec::as_slice)
-            .unwrap_or(&[])
-    }
-
-    pub fn plugin_display_names_for_mcp_server_name(&self, server_name: &str) -> &[String] {
-        self.plugin_display_names_by_mcp_server_name
-            .get(server_name)
-            .map(Vec::as_slice)
-            .unwrap_or(&[])
-    }
-}
-
 #[derive(Debug, Clone, PartialEq)]
 pub struct PluginLoadOutcome {
     plugins: Vec<LoadedPlugin>,
     capability_summaries: Vec<PluginCapabilitySummary>,
-    tool_plugin_provenance: ToolPluginProvenance,
 }
 
 impl Default for PluginLoadOutcome {
@@ -141,42 +118,9 @@ impl PluginLoadOutcome {
             .iter()
             .filter_map(PluginCapabilitySummary::from_plugin)
             .collect::<Vec<_>>();
-        let mut tool_plugin_provenance = ToolPluginProvenance::default();
-        for plugin in &capability_summaries {
-            for connector_id in &plugin.app_connector_ids {
-                tool_plugin_provenance
-                    .plugin_display_names_by_connector_id
-                    .entry(connector_id.0.clone())
-                    .or_default()
-                    .push(plugin.display_name.clone());
-            }
-
-            for server_name in &plugin.mcp_server_names {
-                tool_plugin_provenance
-                    .plugin_display_names_by_mcp_server_name
-                    .entry(server_name.clone())
-                    .or_default()
-                    .push(plugin.display_name.clone());
-            }
-        }
-
-        for plugin_names in tool_plugin_provenance
-            .plugin_display_names_by_connector_id
-            .values_mut()
-            .chain(
-                tool_plugin_provenance
-                    .plugin_display_names_by_mcp_server_name
-                    .values_mut(),
-            )
-        {
-            plugin_names.sort_unstable();
-            plugin_names.dedup();
-        }
-
         Self {
             plugins,
             capability_summaries,
-            tool_plugin_provenance,
         }
     }
 
@@ -221,10 +165,6 @@ impl PluginLoadOutcome {
 
     pub fn capability_summaries(&self) -> &[PluginCapabilitySummary] {
         &self.capability_summaries
-    }
-
-    pub fn tool_plugin_provenance(&self) -> &ToolPluginProvenance {
-        &self.tool_plugin_provenance
     }
 
     pub fn plugins(&self) -> &[LoadedPlugin] {
@@ -948,19 +888,6 @@ mod tests {
                 AppConnectorId("connector_gmail".to_string()),
             ]
         );
-        assert_eq!(
-            outcome.tool_plugin_provenance(),
-            &ToolPluginProvenance {
-                plugin_display_names_by_connector_id: HashMap::from([
-                    (
-                        "connector_example".to_string(),
-                        vec!["plugin-a".to_string(), "plugin-b".to_string()],
-                    ),
-                    ("connector_gmail".to_string(), vec!["plugin-b".to_string()],),
-                ]),
-                plugin_display_names_by_mcp_server_name: HashMap::new(),
-            }
-        );
     }
 
     #[test]
@@ -1049,25 +976,6 @@ mod tests {
                     ..summary("beta@test", "beta-plugin")
                 },
             ]
-        );
-        assert_eq!(
-            outcome.tool_plugin_provenance(),
-            &ToolPluginProvenance {
-                plugin_display_names_by_connector_id: HashMap::from([
-                    (
-                        "connector_example".to_string(),
-                        vec!["alpha-plugin".to_string(), "beta-plugin".to_string()],
-                    ),
-                    (
-                        "connector_gmail".to_string(),
-                        vec!["beta-plugin".to_string()],
-                    ),
-                ]),
-                plugin_display_names_by_mcp_server_name: HashMap::from([
-                    ("alpha".to_string(), vec!["alpha-plugin".to_string()]),
-                    ("beta".to_string(), vec!["beta-plugin".to_string()]),
-                ]),
-            }
         );
     }
